fl16: Add fwupd BOS descriptor
This allows fwupd to automatically detect that this device supports the rp_pico
protocol. So we don't have to upstream a quirk to fwupd and wait for
distributions to update their package.
It will work with fwupd as old as 2.0.2 but is not present on the 1.9.X maintenance branch.

The quirk is at keyboards/framework/fwupd.quirk and the binary data is generated with:
`contrib/generate-ds20.py fwupd.quirk`
It strips away the vid/pid because that is determined by USB enumeration.
That means the same quirk can be used for all our keyboards.

See:

- https://fwupd.github.io/libfwupdplugin/ds20.html
- https://github.com/fwupd/fwupd/tree/main/plugins/rp-pico

Signed-off-by: Daniel Schaefer <dhs@frame.work>